Add GetValidatedConfigurationSection<T>()
Extract some of the logic out of the existing code to provide a
simple method to fetch a section of the configuration, bind it
to a C# POCO, then run validation against that using the
DataAnnotation attributes on the class properties.

This approach provides a middle ground between using IOptions<T>
registered in the DI container and doing weird things like using
AddEagerlyValidatedSettings() which is not strictly compatible
with IOptionsMonitor<T> needs.

Many times in the application start-up process, we need simple
access to the configuration values, but without dealing with
registration as an IOptions<T> object.